app/streamlit-app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachecontrol==0.14.3
    # via firebase-admin
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   nltk
    #   streamlit
cryptography==45.0.4
    # via pyjwt
distro==1.9.0
    # via openai
firebase-admin==6.9.0
    # via streamlit-app (pyproject.toml)
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-api-core==2.25.0
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.172.0
    # via firebase-admin
google-auth==2.40.3
    # via
    #   streamlit-app (pyproject.toml)
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-core==2.4.3
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.21.0
    # via
    #   streamlit-app (pyproject.toml)
    #   firebase-admin
google-cloud-storage==3.1.0
    # via firebase-admin
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.73.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.73.0
    # via google-api-core
h11==0.16.0
    # via httpcore
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   firebase-admin
    #   langsmith
    #   openai
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via nltk
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.3.25
    # via streamlit-app (pyproject.toml)
langchain-core==0.3.65
    # via
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.22
    # via streamlit-app (pyproject.toml)
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.45
    # via
    #   langchain
    #   langchain-core
lxml==5.4.0
    # via pytrends
markupsafe==3.0.2
    # via jinja2
msgpack==1.1.0
    # via cachecontrol
narwhals==1.42.0
    # via altair
nltk==3.9.1
    # via streamlit-app (pyproject.toml)
numpy==2.3.0
    # via
    #   pandas
    #   pydeck
    #   streamlit
openai==1.86.0
    # via langchain-openai
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   langchain-core
    #   langsmith
    #   streamlit
pandas==2.3.0
    # via
    #   pytrends
    #   streamlit
pillow==11.2.1
    # via streamlit
praw==7.8.1
    # via streamlit-app (pyproject.toml)
prawcore==2.4.0
    # via praw
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==20.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyjwt==2.10.1
    # via firebase-admin
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via streamlit-app (pyproject.toml)
pytrends==4.9.2
    # via streamlit-app (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   nltk
    #   tiktoken
requests==2.32.4
    # via
    #   streamlit-app (pyproject.toml)
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
    #   langchain
    #   langsmith
    #   prawcore
    #   pytrends
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
    #   update-checker
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via langchain
streamlit==1.45.1
    # via streamlit-app (pyproject.toml)
tenacity==9.1.2
    # via
    #   langchain-core
    #   streamlit
tiktoken==0.9.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via
    #   nltk
    #   openai
typing-extensions==4.14.0
    # via
    #   altair
    #   anyio
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
update-checker==0.18.0
    # via praw
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.4.0
    # via requests
websocket-client==1.8.0
    # via praw
yt-dlp==2025.6.9
    # via streamlit-app (pyproject.toml)
zstandard==0.23.0
    # via langsmith
